+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## in develop
 
 * Added sub owner filter.
+* Added extraction of sub owner for sysOwner field.
 
 ## 1.3.2